# apps/webapi/public/openapi.yaml
openapi: 3.0.0
info:
  title: (title)
  version: 0.0.0
...
paths:
...
  /organization/{organizationId}/invitation:
    post:
      operationId: Invitation_create
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
              required:
                - name
                - email
      security:
        - BearerAuth: []
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer